rename `blockdata` module to `primitives`
