+++ fuzz/fuzz_targets/deserialize_block.rs
@@ -1,7 +1,7 @@
 extern crate bitcoin;
 
 fn do_test(data: &[u8]) {
-    let _: Result<bitcoin::blockdata::block::Block, _>= bitcoin::consensus::encode::deserialize(data);
+    let _: Result<bitcoin::Block, _>= bitcoin::consensus::encode::deserialize(data);
 }
 
 #[cfg(feature = "afl")]
