+++ src/lib.rs
@@ -60,20 +60,20 @@ mod test_macros;
 mod internal_macros;
 #[macro_use]
 pub mod network;
-pub mod blockdata;
+pub mod primitives;
 pub mod util;
 pub mod consensus;
 pub mod hash_types;
 
 pub use hash_types::*;
-pub use blockdata::block::Block;
-pub use blockdata::block::BlockHeader;
-pub use blockdata::script::Script;
-pub use blockdata::transaction::Transaction;
-pub use blockdata::transaction::TxIn;
-pub use blockdata::transaction::TxOut;
-pub use blockdata::transaction::OutPoint;
-pub use blockdata::transaction::SigHashType;
+pub use primitives::block::Block;
+pub use primitives::block::BlockHeader;
+pub use primitives::script::Script;
+pub use primitives::transaction::Transaction;
+pub use primitives::transaction::TxIn;
+pub use primitives::transaction::TxOut;
+pub use primitives::transaction::OutPoint;
+pub use primitives::transaction::SigHashType;
 pub use consensus::encode::VarInt;
 pub use network::constants::Network;
 pub use util::Error;
