+++ src/primitives/transaction.rs
@@ -30,9 +30,9 @@ use hashes::{self, Hash, sha256d};
 use hashes::hex::FromHex;
 
 use util::endian;
-use blockdata::constants::WITNESS_SCALE_FACTOR;
-#[cfg(feature="bitcoinconsensus")] use blockdata::script;
-use blockdata::script::Script;
+use primitives::constants::WITNESS_SCALE_FACTOR;
+#[cfg(feature="bitcoinconsensus")] use primitives::script;
+use primitives::script::Script;
 use consensus::{encode, Decodable, Encodable};
 use hash_types::*;
 use VarInt;
@@ -74,7 +74,7 @@ impl OutPoint {
     /// # Examples
     ///
     /// ```rust
-    /// use bitcoin::blockdata::constants::genesis_block;
+    /// use bitcoin::primitives::constants::genesis_block;
     /// use bitcoin::network::constants::Network;
     ///
     /// let block = genesis_block(Network::Bitcoin);
@@ -687,8 +687,8 @@ mod tests {
     use super::{OutPoint, ParseOutPointError, Transaction, TxIn};
 
     use std::str::FromStr;
-    use blockdata::constants::WITNESS_SCALE_FACTOR;
-    use blockdata::script::Script;
+    use primitives::constants::WITNESS_SCALE_FACTOR;
+    use primitives::script::Script;
     use consensus::encode::serialize;
     use consensus::encode::deserialize;
 
@@ -750,7 +750,7 @@ mod tests {
     #[test]
     fn test_is_coinbase () {
         use network::constants::Network;
-        use blockdata::constants;
+        use primitives::constants;
 
         let genesis = constants::genesis_block(Network::Bitcoin);
         assert! (genesis.txdata[0].is_coin_base());
@@ -1278,7 +1278,7 @@ mod tests {
     fn test_transaction_verify () {
         use hashes::hex::FromHex;
         use std::collections::HashMap;
-        use blockdata::script;
+        use primitives::script;
         // a random recent segwit transaction from blockchain using both old and segwit inputs
         let mut spending: Transaction = deserialize(Vec::from_hex("020000000001031cfbc8f54fbfa4a33a30068841371f80dbfe166211242213188428f437445c91000000006a47304402206fbcec8d2d2e740d824d3d36cc345b37d9f65d665a99f5bd5c9e8d42270a03a8022013959632492332200c2908459547bf8dbf97c65ab1a28dec377d6f1d41d3d63e012103d7279dfb90ce17fe139ba60a7c41ddf605b25e1c07a4ddcb9dfef4e7d6710f48feffffff476222484f5e35b3f0e43f65fc76e21d8be7818dd6a989c160b1e5039b7835fc00000000171600140914414d3c94af70ac7e25407b0689e0baa10c77feffffffa83d954a62568bbc99cc644c62eb7383d7c2a2563041a0aeb891a6a4055895570000000017160014795d04cc2d4f31480d9a3710993fbd80d04301dffeffffff06fef72f000000000017a91476fd7035cd26f1a32a5ab979e056713aac25796887a5000f00000000001976a914b8332d502a529571c6af4be66399cd33379071c588ac3fda0500000000001976a914fc1d692f8de10ae33295f090bea5fe49527d975c88ac522e1b00000000001976a914808406b54d1044c429ac54c0e189b0d8061667e088ac6eb68501000000001976a914dfab6085f3a8fb3e6710206a5a959313c5618f4d88acbba20000000000001976a914eb3026552d7e3f3073457d0bee5d4757de48160d88ac0002483045022100bee24b63212939d33d513e767bc79300051f7a0d433c3fcf1e0e3bf03b9eb1d70220588dc45a9ce3a939103b4459ce47500b64e23ab118dfc03c9caa7d6bfc32b9c601210354fd80328da0f9ae6eef2b3a81f74f9a6f66761fadf96f1d1d22b1fd6845876402483045022100e29c7e3a5efc10da6269e5fc20b6a1cb8beb92130cc52c67e46ef40aaa5cac5f0220644dd1b049727d991aece98a105563416e10a5ac4221abac7d16931842d5c322012103960b87412d6e169f30e12106bdf70122aabb9eb61f455518322a18b920a4dfa887d30700")
             .unwrap().as_slice()).unwrap();
