+++ src/util/bip158.rs
@@ -51,16 +51,11 @@ use std::fmt::{Display, Formatter};
 use std::io::Cursor;
 use std::cmp::Ordering;
 
-
 use hashes::{Hash, siphash24};
-use hash_types::{BlockHash, FilterHash, FilterHeader};
 
-use blockdata::block::Block;
-use blockdata::script::Script;
-use blockdata::transaction::OutPoint;
 use consensus::{Decodable, Encodable};
-use consensus::encode::VarInt;
 use util::endian;
+use {Block, BlockHash, FilterHash, FilterHeader, OutPoint, Script, VarInt};
 
 /// Golomb encoding parameter as in BIP-158, see also https://gist.github.com/sipa/576d5f09c3b86c3b1b75598d799fc845
 const P: u8 = 19;
