+++ src/util/contracthash.rs
@@ -24,7 +24,7 @@ use secp256k1::{self, Secp256k1};
 use PrivateKey;
 use PublicKey;
 use hashes::{sha256, Hash, HashEngine, Hmac, HmacEngine};
-use blockdata::{opcodes, script};
+use primitives::{opcodes, script};
 
 use std::{error, fmt};
 
@@ -280,11 +280,8 @@ mod tests {
     use secp256k1::rand::thread_rng;
     use std::str::FromStr;
 
-    use blockdata::script::Script;
-    use network::constants::Network;
-
+    use {Network, PublicKey, Script};
     use super::*;
-    use PublicKey;
 
     macro_rules! hex (($hex:expr) => (Vec::from_hex($hex).unwrap()));
     macro_rules! hex_key (($hex:expr) => (PublicKey::from_slice(&hex!($hex)).unwrap()));
