+++ src/util/psbt/error.rs
@@ -15,8 +15,8 @@
 use std::error;
 use std::fmt;
 
-use blockdata::transaction::Transaction;
 use util::psbt::raw;
+use Transaction;
 
 use hashes;
