+++ src/util/psbt/map/global.rs
@@ -16,12 +16,12 @@ use std::collections::BTreeMap;
 use std::collections::btree_map::Entry;
 use std::io::{self, Cursor};
 
-use blockdata::transaction::Transaction;
 use consensus::{encode, Encodable, Decodable};
 use util::psbt::map::Map;
 use util::psbt::raw;
 use util::psbt;
 use util::psbt::Error;
+use Transaction;
 
 /// A key-value map for global data.
 #[derive(Clone, Debug, PartialEq)]
