+++ src/util/psbt/serialize.rs
@@ -19,13 +19,11 @@
 
 use std::io;
 
-use blockdata::script::Script;
-use blockdata::transaction::{SigHashType, Transaction, TxOut};
 use consensus::encode::{self, serialize, Decodable};
 use util::bip32::{ChildNumber, Fingerprint, KeySource};
 use hashes::{hash160, ripemd160, sha256, sha256d, Hash};
-use util::key::PublicKey;
 use util::psbt;
+use {PublicKey, Script, SigHashType, Transaction, TxOut};
 
 /// A trait for serializing a value as raw data for insertion into PSBT
 /// key-value pairs.
